fix(#5751): shrink the directory walk atom to a single-level listing
